refactoring/#17: categoryswiper 뷰로직 비지니스로직 분리
